Last fixes to generic analysis PR
* Move `GetRelativePathTo` from Transforms to Analysis namespace
* Add `TryFindBestRootDirectory` helper extension method
* Improve codestyle a bit (ReSharper)
