+++ Project2015To2017/Transforms/Helpers.cs
@@ -19,24 +19,5 @@ namespace Project2015To2017.Transforms
 		{
 			return source.Elements().Where(e => e.Name.LocalName == localName);
 		}
-		
-		public static string GetRelativePathTo(this FileSystemInfo from, FileSystemInfo to)
-		{
-			string GetPath(FileSystemInfo fsi)
-			{
-				return (fsi is DirectoryInfo d) ? (d.FullName.TrimEnd('\\') + "\\") : fsi.FullName;
-			}
-
-			var fromPath = GetPath(from);
-			var toPath = GetPath(to);
-
-			var fromUri = new Uri(fromPath);
-			var toUri = new Uri(toPath);
-
-			var relativeUri = fromUri.MakeRelativeUri(toUri);
-			var relativePath = Uri.UnescapeDataString(relativeUri.ToString());
-
-			return relativePath.Replace('/', Path.DirectorySeparatorChar);
-		}
 	}
 }
